+++ spec/request_validation/parameter_validation_spec.rb
@@ -96,6 +96,11 @@ RSpec.describe 'Parameter validation' do
 
     it 'adds filtered query parameters to env ' do
       get '/search', params
+      expect(last_request.env[OpenapiFirst::PARAMS]).to eq expected_params
+    end
+
+    it 'still adds PARAMETERS to env ' do
+      get '/search', params
       expect(last_request.env[OpenapiFirst::PARAMETERS]).to eq expected_params
     end
 
@@ -121,7 +126,7 @@ RSpec.describe 'Parameter validation' do
       get '/search', params.merge(foo: 'bar')
 
       expect(last_response.status).to eq 200
-      expect(last_request.env[OpenapiFirst::PARAMETERS]).to eq expected_params
+      expect(last_request.env[OpenapiFirst::PARAMS]).to eq expected_params
     end
 
     describe 'with array query parameters' do
@@ -135,7 +140,7 @@ RSpec.describe 'Parameter validation' do
           }
           get '/default-style', params
           expect(last_response.status).to eq(200), last_response.body
-          parsed_parameters = last_request.env[OpenapiFirst::PARAMETERS]
+          parsed_parameters = last_request.env[OpenapiFirst::PARAMS]
           expect(parsed_parameters['strings']).to eq %w[a b c]
           expect(parsed_parameters['integers']).to eq [2, 3, 4]
         end
@@ -143,14 +148,14 @@ RSpec.describe 'Parameter validation' do
         it 'parses nested array' do
           get '/default-style?nested[integers]=2,3,4'
           expect(last_response.status).to eq(200), last_response.body
-          parsed_parameters = last_request.env[OpenapiFirst::PARAMETERS]
+          parsed_parameters = last_request.env[OpenapiFirst::PARAMS]
           expect(parsed_parameters['nested[integers]']).to eq [2, 3, 4]
         end
 
         it 'ignores empty query params' do
           get '/default-style'
           expect(last_response.status).to eq(200), last_response.body
-          parsed_parameters = last_request.env[OpenapiFirst::PARAMETERS]
+          parsed_parameters = last_request.env[OpenapiFirst::PARAMS]
           expect(parsed_parameters['strings']).to be_nil
         end
 
@@ -180,7 +185,7 @@ RSpec.describe 'Parameter validation' do
         params = {}
         get '/with-default-query-param', params
         expect(last_response.status).to eq(200)
-        parsed_parameters = last_request.env[OpenapiFirst::PARAMETERS]
+        parsed_parameters = last_request.env[OpenapiFirst::PARAMS]
         expect(parsed_parameters['has_default']).to eq true
       end
 
@@ -196,7 +201,7 @@ RSpec.describe 'Parameter validation' do
         params = { has_default: false }
         get '/with-default-query-param', params
         expect(last_response.status).to eq(200)
-        parsed_parameters = last_request.env[OpenapiFirst::PARAMETERS]
+        parsed_parameters = last_request.env[OpenapiFirst::PARAMS]
         expect(parsed_parameters['has_default']).to eq false
       end
     end
@@ -282,7 +287,7 @@ RSpec.describe 'Parameter validation' do
 
     describe 'type conversion' do
       def last_params
-        last_request.env[OpenapiFirst::PARAMETERS]
+        last_request.env[OpenapiFirst::PARAMS]
       end
 
       it 'converts to integer' do